32-ый commit 16-ый спринт, работа в MVC
