+++ app/src/main/java/com/example/playlistmaker/Search/domain/api/ConsumerData.kt
@@ -2,6 +2,6 @@ package com.example.playlistmaker.Search.domain.api
 
 sealed interface ConsumerData<T> {
     data class Data<T>(val data: T) : ConsumerData<T>
-    data class Error<T>(val message: String) : ConsumerData<T>
+    data class Error<T>(val errorMessage: String) : ConsumerData<T>
 }
